heavy refactor of tacacs agent
