--- libraries/tacacsrs_agent/src/operations/mod.rs
@@ -1,10 +1,0 @@
-//! Typed operation execution entry points.
-//!
-//! Accounting and authorization share the same routing and failover machinery,
-//! but each operation has its own request/response types and upstream send
-//! method. This module keeps those typed adapters together so adding another
-//! TACACS+ operation has an obvious home.
-
-mod accounting;
-mod authorization;
-mod routed;
