+++ libraries/tacacsrs_agent/src/services/client_api/listener/tcp.rs
@@ -1,21 +1,19 @@
-//! Loopback TCP IPC listener.
+//! Loopback TCP IPC listener for non-Unix builds.
 
 use std::net::SocketAddr;
-use std::sync::Arc;
 
 use anyhow::{Context, bail};
 use tacacsrs_agent_client::ipc::tacacs_agent_server::TacacsAgentServer;
 use tokio_stream::wrappers::TcpListenerStream;
 
-use crate::ipc::GrpcService;
-use crate::routing::RoutingState;
 use crate::runtime::shutdown_signal;
+use crate::services::client_api::ClientApiService;
 
 /// Serves loopback TCP IPC clients until shutdown is requested.
 ///
 /// This path exists primarily for non-Unix development workflows where a Unix
 /// domain socket is not available.
-pub(crate) async fn serve(address: SocketAddr, state: Arc<RoutingState>) -> anyhow::Result<()> {
+pub(crate) async fn serve(address: SocketAddr, service: ClientApiService) -> anyhow::Result<()> {
     if !address.ip().is_loopback() {
         log::error!("Refusing non-loopback TCP IPC endpoint: {address}");
         bail!("TCP IPC endpoint must be loopback-only: {address}");
@@ -25,17 +23,16 @@ pub(crate) async fn serve(address: SocketAddr, state: Arc<RoutingState>) -> anyh
         .await
         .with_context(|| format!("Failed to bind TCP IPC endpoint {address}"))?;
     let incoming = TcpListenerStream::new(listener);
-    let grpc_service = GrpcService::new(Arc::clone(&state));
 
     log::info!("Listening for IPC clients on TCP {address}");
 
     tonic::transport::Server::builder()
-        .add_service(TacacsAgentServer::new(grpc_service))
+        .add_service(TacacsAgentServer::new(service.grpc_service()))
         .serve_with_incoming_shutdown(incoming, shutdown_signal())
         .await
         .with_context(|| format!("TCP IPC server {address} failed"))?;
 
     log::info!("Shutdown signal received; draining active IPC clients");
-    state.wait_for_active_clients().await;
+    service.wait_for_active_requests().await;
     Ok(())
 }
