+++ libraries/tacacsrs_agent/src/services/client_api/upstream_bridge/accounting.rs
@@ -1,10 +1,11 @@
-//! Accounting request execution for [`RoutingState`].
+//! Accounting request execution for [`UpstreamBridge`].
 
 use async_trait::async_trait;
 use tacacsrs_agent_client::{AccountingOperation, AccountingOperationResponse, ServiceError};
 
+use super::mapping::{build_accounting_request, to_accounting_response};
+use super::UpstreamBridge;
 use super::routed::RoutedOperation;
-use crate::routing::RoutingState;
 use crate::upstream::UpstreamConnection;
 
 struct AccountingRoute;
@@ -21,17 +22,20 @@ impl RoutedOperation for AccountingRoute {
         connection: &dyn UpstreamConnection,
         request: &Self::Request,
     ) -> anyhow::Result<Self::Response> {
-        connection.send_accounting(request).await
+        let reply = connection
+            .send_accounting(build_accounting_request(request))
+            .await?;
+        Ok(to_accounting_response(connection.server_address(), reply))
     }
 }
 
-impl RoutingState {
+impl UpstreamBridge {
     /// Executes one IPC accounting RPC against the currently selected upstream
     /// TACACS+ server.
     ///
     /// Connection reuse and single-connection negotiation are handled by the
     /// networking layer behind the selected upstream connection.
-    pub(crate) async fn execute_accounting_request(
+    pub(in crate::services::client_api) async fn execute_accounting_request(
         &self,
         request: AccountingOperation,
     ) -> Result<AccountingOperationResponse, ServiceError> {
