+++ libraries/tacacsrs_agent/src/services/client_api/upstream_bridge/routed.rs
@@ -1,23 +1,22 @@
-//! Shared operation helpers for [`RoutingState`].
+//! Shared operation helpers for [`UpstreamBridge`].
 
 use async_trait::async_trait;
 use tacacsrs_agent_client::ServiceError;
 
-use crate::routing::RoutingState;
+use super::UpstreamBridge;
 use crate::upstream::UpstreamConnection;
 
-/// Operation-specific hooks used by the shared TACACS+ routing state machine.
+/// Operation-specific hooks used by the client API upstream bridge.
 ///
 /// Accounting and authorization differ in request/response types and upstream
-/// send methods, but they share the same client tracking, active-server
-/// selection, failure recording, and failover behavior. Implementing this trait
-/// keeps those specializations small while preserving typed request and response
-/// values.
+/// send methods, but they share the same active-server selection, failure
+/// recording, and failover behavior. Implementing this trait keeps those
+/// specializations small while preserving typed request and response values.
 #[async_trait]
-pub(crate) trait RoutedOperation: Send + Sync + 'static {
-    /// IPC-level request type for the operation.
+pub(super) trait RoutedOperation: Send + Sync + 'static {
+    /// Client API request type for the operation.
     type Request: Send + Sync;
-    /// IPC-level response type for the operation.
+    /// Client API response type for the operation.
     type Response: Send;
 
     /// Lowercase operation name for log messages.
@@ -32,25 +31,29 @@ pub(crate) trait RoutedOperation: Send + Sync + 'static {
     ) -> anyhow::Result<Self::Response>;
 }
 
-impl RoutingState {
-    /// Executes one IPC operation against the currently selected upstream
-    /// TACACS+ server.
+impl UpstreamBridge {
+    /// Executes one client API operation against the currently selected
+    /// upstream TACACS+ server.
     ///
     /// The networking layer owns dedicated versus single-connection behavior.
-    /// `RoutingState` only selects a configured server, records failures, and
-    /// advances failover state for subsequent IPC requests.
-    pub(crate) async fn execute_operation<Operation>(
+    /// `UpstreamBridge` asks the upstream manager to select a configured
+    /// server, records failures, and advances failover state for subsequent
+    /// local client API requests.
+    pub(super) async fn execute_operation<Operation>(
         &self,
         request: Operation::Request,
     ) -> Result<Operation::Response, ServiceError>
     where
         Operation: RoutedOperation,
     {
-        let _client_guard = self.start_client_request();
-        let bound_server = self.bind_server_for_new_session().await.map_err(|error| {
-            log::warn!("Failed to bind IPC request to an upstream server: {error:#}");
-            ServiceError::new(error.to_string()).retriable(true)
-        })?;
+        let bound_server = self
+            .upstream_manager
+            .bind_server_for_new_session()
+            .await
+            .map_err(|error| {
+                log::warn!("Failed to bind client API request to an upstream server: {error:#}");
+                ServiceError::new(error.to_string()).retriable(true)
+            })?;
 
         log::debug!(
             "Executing {} request via {} (server index {})",
@@ -67,7 +70,9 @@ impl RoutingState {
                     Operation::DISPLAY_NAME,
                     bound_server.connection.server_address(),
                 );
-                self.note_bound_server_failure(&bound_server).await;
+                self.upstream_manager
+                    .note_bound_server_failure(&bound_server)
+                    .await;
                 Err(ServiceError::new(error.to_string())
                     .with_server(bound_server.connection.server_address())
                     .retriable(true))
