+++ libraries/tacacsrs_agent/src/services/tacacs_proxy/upstream_bridge/error.rs
@@ -1,0 +1,7 @@
+//! Error boundary for one proxied TACACS+ connection.
+
+#[derive(Debug)]
+pub(super) enum ProxyConnectionError {
+    Downstream(anyhow::Error),
+    Upstream(anyhow::Error),
+}
